+++ src/main/groovy/com/cedarsoftware/ncube/DeltaProcessor.groovy
@@ -1,5 +1,6 @@
 package com.cedarsoftware.ncube
 
+import com.cedarsoftware.ncube.formatters.NCubeTestReader
 import com.cedarsoftware.ncube.util.LongHashSet
 import com.cedarsoftware.util.CaseInsensitiveMap
 import com.cedarsoftware.util.CaseInsensitiveSet
@@ -34,6 +35,7 @@ class DeltaProcessor
     public static final String DELTA_CELLS = 'delta-cel'
     public static final String DELTA_AXES_COLUMNS = 'delta-col'
     public static final String DELTA_AXES = 'delta-axis'
+    public static final String DELTA_TESTS = 'delta-test'
 
     public static final String DELTA_NCUBE_META_PUT = 'ncube-meta-put'
     public static final String DELTA_NCUBE_META_REMOVE = 'ncube-meta-del'
@@ -90,6 +92,9 @@ class DeltaProcessor
             return null
         }
 
+        delta[DELTA_TESTS] = getTestDeltas(baseCube, changeCube)
+        baseCube.removeMetaProperty(NCube.METAPROPERTY_TEST_DATA)
+        changeCube.removeMetaProperty(NCube.METAPROPERTY_TEST_DATA)
 
         Map<String, Map<String, Object>> metaDeltaMap = new CaseInsensitiveMap<>()
         delta[DELTA_NCUBE] = metaDeltaMap
@@ -283,7 +288,8 @@ class DeltaProcessor
         return areNCubeDifferencesOK(branchDelta, headDelta) &&
                 areAxisDifferencesOK(branchDelta, headDelta) &&
                 areColumnDifferencesOK(branchDelta, headDelta) &&
-                areCellDifferencesOK(branchDelta, headDelta)
+                areCellDifferencesOK(branchDelta, headDelta) &&
+                areTestDifferencesOK(branchDelta, headDelta)
     }
 
     /**
@@ -465,6 +471,38 @@ class DeltaProcessor
                 CellInfo info2 = new CellInfo(biggerChangeSet[deltaCoord])
 
                 if (info1 != info2)
+                {
+                    return false
+                }
+            }
+        }
+        return true
+    }
+
+    private static boolean areTestDifferencesOK(Map<String, Object> branchDelta, Map<String, Object> headDelta)
+    {
+        Map<String, Map<String, Delta>> deltaMap1 = branchDelta[DELTA_TESTS] as Map
+        Map<String, Map<String, Delta>> deltaMap2 = headDelta[DELTA_TESTS] as Map
+
+        for (Map.Entry<String, Map<String, Delta>> entry1 : deltaMap1.entrySet())
+        {
+            String testName = entry1.key
+            if (!deltaMap2.containsKey(testName))
+            {
+                continue // no test changed with same ID
+            }
+
+            Map<String, Delta> changes2 = deltaMap2[testName]
+            for (Map.Entry<String, Delta> changes1 : entry1.value)
+            {
+                String objName = changes1.key
+                if (!changes2.containsKey(objName)) {
+                    continue // no coord or assertion changes with same ID
+                }
+
+                Delta delta1 = changes1.value
+                Delta delta2 = changes2[objName]
+                if (!DeepEquals.deepEquals(delta1.destVal, delta2.destVal))
                 {
                     return false
                 }
@@ -656,6 +694,134 @@ class DeltaProcessor
         return delta
     }
 
+    private static List<Delta> getTestDeltaList(NCube newCube, NCube oldCube)
+    {
+        List<Delta> deltas = []
+        Map<String, Map<String, Delta>> deltaMap = getTestDeltas(newCube, oldCube)
+        for (Map.Entry<String, Map<String, Delta>> testDeltaEntry : deltaMap)
+        {
+            Map<String, Delta> objMap = testDeltaEntry.value
+            for (Map.Entry<String, Delta> objDeltaEntry : objMap)
+            {
+                Delta delta = objDeltaEntry.value
+                deltas.add(delta)
+            }
+        }
+        return deltas
+    }
+
+    private static Map<String, Map<String, Delta>> getTestDeltas(NCube newCube, NCube oldCube)
+    {
+        List<NCubeTest> newCubeTests = newCube.testData
+        List<NCubeTest> oldCubeTests = oldCube.testData
+        Map<String, Map<String, Delta>> deltaMap = new CaseInsensitiveMap()
+        if (!newCubeTests && !oldCubeTests)
+        {
+            return deltaMap
+        }
+
+        Object[] newCubeTestObj = newCubeTests.toArray()
+        Object[] oldCubeTestObj = oldCubeTests.toArray()
+
+        for (NCubeTest newCubeTest : newCubeTests)
+        {
+            String newCubeTestName = newCubeTest.name
+            NCubeTest oldCubeTest = oldCubeTests.find { NCubeTest test -> test.name == newCubeTestName }
+
+            if (oldCubeTest)
+            {   // has the same test
+                Map<String, Delta> testDeltaMap = new CaseInsensitiveMap()
+
+                // coords
+                Map<String, CellInfo> oldTestCoords = oldCubeTest.coord
+                for (Map.Entry<String, CellInfo> newTestCoordEntry : newCubeTest.coord)
+                {
+                    String key = newTestCoordEntry.key
+                    CellInfo newTestCoord = newTestCoordEntry.value
+                    MapEntry newEntry = new MapEntry(key, newTestCoord)
+                    if (oldTestCoords.containsKey(key))
+                    {   // has the same coord name
+                        CellInfo oldTestCoord = oldTestCoords[key]
+                        oldTestCoords.remove(key)
+                        if (!DeepEquals.deepEquals(newTestCoord, oldTestCoord))
+                        {   // value changed
+                            String s = "Change ${newCubeTestName} coord ${key}: ${oldTestCoord.value} ==> ${newTestCoord.value}"
+                            MapEntry oldEntry = new MapEntry(key, oldTestCoord)
+                            testDeltaMap[key] = new Delta(Delta.Location.TEST_COORD, Delta.Type.UPDATE, s, newCubeTestName, oldEntry, newEntry, oldCubeTestObj, newCubeTestObj)
+                        }
+                    }
+                    else
+                    {   // coord no longer present
+                        String s = "Add ${newCubeTestName} coord {${key}: ${newTestCoord.value}}"
+                        testDeltaMap[key] = new Delta(Delta.Location.TEST_COORD, Delta.Type.ADD, s, newCubeTestName, null, newEntry, oldCubeTestObj, newCubeTestObj)
+                    }
+                }
+                for (Map.Entry<String, CellInfo> oldTestCoordEntry : oldTestCoords)
+                {   // coord added
+                    String key = oldTestCoordEntry.key
+                    CellInfo value = oldTestCoordEntry.value
+                    String s = "Remove ${newCubeTest.name} coord {${key}: ${value.value}}"
+                    MapEntry oldEntry = new MapEntry(key, value)
+                    testDeltaMap[key] = new Delta(Delta.Location.TEST_COORD, Delta.Type.DELETE, s, newCubeTestName, oldEntry, null, oldCubeTestObj, newCubeTestObj)
+                }
+
+                // assertions
+                CellInfo[] oldCubeAsserts = oldCubeTest.assertions
+                Set<String> checkedAsserts = []
+                for (CellInfo newAssert : newCubeTest.assertions)
+                {
+                    String newVal = newAssert.value
+                    CellInfo oldAssert = oldCubeAsserts.find { CellInfo oldAssert ->
+                        oldAssert.value == newVal
+                    }
+                    if (oldAssert)
+                    {
+                        checkedAsserts.add(newVal)
+                        if (!(oldAssert.isUrl == newAssert.isUrl))
+                        {
+                            String s = "Change assertion ${newVal} ${oldAssert.isUrl ? 'URL' : 'Value'} ==> ${newAssert.isUrl ? 'URL' : 'Value'}"
+                            testDeltaMap[newVal] = new Delta(Delta.Location.TEST_ASSERT, Delta.Type.UPDATE, s, newCubeTestName, new CellInfo(oldAssert), new CellInfo(newAssert), oldCubeTestObj, newCubeTestObj)
+                        }
+                    }
+                    else
+                    {
+                        String s = "Add assertion ${newVal}"
+                        testDeltaMap[newVal] = new Delta(Delta.Location.TEST_ASSERT, Delta.Type.ADD, s, newCubeTestName, null, new CellInfo(newAssert), oldCubeTestObj, newCubeTestObj)
+                    }
+                }
+                for (CellInfo oldAssert : oldCubeTest.assertions)
+                {
+                    String oldVal = oldAssert.value
+                    if (!checkedAsserts.contains(oldVal))
+                    {
+                        String s = "Remove assertion ${oldVal}"
+                        testDeltaMap[oldVal] = new Delta(Delta.Location.TEST_ASSERT, Delta.Type.DELETE, s, newCubeTestName, new CellInfo(oldAssert), null, oldCubeTestObj, newCubeTestObj)
+                    }
+                }
+
+                oldCubeTests.remove(oldCubeTest) // remove to have shortened list
+                if (!testDeltaMap.empty)
+                {
+                    deltaMap[newCubeTestName] = testDeltaMap
+                }
+            }
+            else
+            {   // added test
+                String s = "Add test ${newCubeTest.name}"
+                deltaMap[newCubeTestName] = [(Delta.Type.ADD.name()): new Delta(Delta.Location.TEST, Delta.Type.ADD, s, newCubeTestName, null, newCubeTest, oldCubeTestObj, newCubeTestObj)]
+            }
+        }
+
+        for (NCubeTest oldCubeTest : oldCubeTests)
+        {   // deleted test
+            String oldCubeTestName = oldCubeTest.name
+            String s = "Remove test ${oldCubeTestName}"
+            deltaMap[oldCubeTestName] = [(Delta.Type.DELETE.name()): new Delta(Delta.Location.TEST, Delta.Type.DELETE, s, oldCubeTestName, oldCubeTest, null, oldCubeTestObj, newCubeTestObj)]
+        }
+
+        return deltaMap
+    }
+
     /**
      * Return a list of Delta objects describing the differences between two n-cubes.
      * @param oldCube NCube to compare 'this' n-cube to
@@ -666,6 +832,10 @@ class DeltaProcessor
     static List<Delta> getDeltaDescription(NCube newCube, NCube oldCube)
     {
         List<Delta> changes = []
+
+        changes.addAll(getTestDeltaList(newCube, oldCube))
+        newCube.removeMetaProperty(NCube.METAPROPERTY_TEST_DATA)
+        oldCube.removeMetaProperty(NCube.METAPROPERTY_TEST_DATA)
 
         getNCubeChanges(newCube, oldCube, changes)