+++ src/main/groovy/com/cedarsoftware/ncube/NCubeReadOnlyPersister.groovy
@@ -136,4 +136,6 @@ interface NCubeReadOnlyPersister
      * @throws IllegalArgumentException if the passed in cubeName does not exist.
      */
     String getTestData(ApplicationID appId, String cubeName, String username)
+
+    String getTestData(Long cubeId, String user)
 }