+++ changelog.md
@@ -1,6 +1,9 @@
 ### Revision History
 * 4.1.0
   * Separated out server.port property. Renamed server.* properties for JsonHttpProxy to target.*.
+  * Added NCubeTest as part of DeltaProcessor.
+  * Added loadCubeWithTests to get both the cube and its tests together.
+  * Removed saveTests from NCubeManager; this is now wrapped into updateCube in the persister.
   * Bug fix: Reference Axes not get checked properly on pull request merge vs straight commit branch.
   * Added appId to a couple InvalidCoordinateExceptions for more clarity in logs
 * 4.0.9